js review changes
