+++ app/webpacker/components/RegistrationsV2/Register/RegistrationClosedMessage.jsx
@@ -2,6 +2,7 @@ import { DateTime } from 'luxon';
 import React, { useEffect, useState } from 'react';
 import { Message } from 'semantic-ui-react';
 import I18nHTMLTranslate from '../../I18nHTMLTranslate';
+import { fullTimeDiff } from '../../../lib/utils/dates';
 
 export default function RegistrationClosedMessage({
   registrationStart,
@@ -17,23 +18,12 @@ export default function RegistrationClosedMessage({
       && timeLeft.seconds === 0) {
       onTimerEnd();
     }
-  }, [onTimerEnd, timeLeft.days, timeLeft.hours, timeLeft.minutes, timeLeft.seconds]);
+  }, [onTimerEnd, timeLeft]);
 
   useEffect(() => {
-    const updateTimer = () => {
-      const now = DateTime.local();
-
-      const diff = start.diff(now, ['days', 'hours', 'minutes', 'seconds']).toObject();
-      setTimeLeft({
-        days: Math.floor(diff.days),
-        hours: Math.floor(diff.hours),
-        minutes: Math.floor(diff.minutes),
-        seconds: Math.floor(diff.seconds),
-      });
-    };
-
-    const intervalId = setInterval(updateTimer, 1000); // Update every second
-    updateTimer(); // Initial call to set the state right away
+    const intervalId = setInterval(() => {
+      setTimeLeft(fullTimeDiff(start));
+    }, 1000); // Update every second
 
     return () => clearInterval(intervalId); // Cleanup interval on unmount
   }, [registrationStart, onTimerEnd, start]);
@@ -42,7 +32,7 @@ export default function RegistrationClosedMessage({
     <Message color="blue">
       { timeLeft.days < 1 && timeLeft.hours < 1
         ? <I18nHTMLTranslate i18nKey="competitions.registration_v2.register.will_open_countdown" options={{ minutes: timeLeft.minutes, seconds: timeLeft.seconds }} />
-        : <I18nHTMLTranslate i18nKey="registrations.will_open_html" options={{ days: start.toRelative(), time: DateTime.fromISO(registrationStart).toLocaleString(DateTime.DATETIME_FULL) }} />}
+        : <I18nHTMLTranslate i18nKey="registrations.will_open_html" options={{ days: start.toRelative(), time: start.toLocaleString(DateTime.DATETIME_FULL) }} />}
     </Message>
   );
 }
